Añadiendo edicion y eliminacion de proyectos. Ajuste en imagenes

+++ app/Http/Controllers/Admin/AdminHostsController.php
@@ -206,7 +206,7 @@ class AdminHostsController extends Controller
             'delete_reasons' => 'required|string|max:500|min:10',
         ]);
 
-        if ($host->host->avatar) {
+        if ($host->host->avatar && $host->host->avatar !== 'logo.svg') {
             $this->imageService->deleteImage($host->host->avatar);
         }
